+++ lib/src/theme/sort_icon_builder.dart
@@ -1,0 +1,6 @@
+import 'package:davi/src/sort_direction.dart';
+import 'package:davi/src/theme/sort_icon_colors.dart';
+import 'package:flutter/widgets.dart';
+
+typedef SortIconBuilder = Widget Function(
+    DaviSortDirection direction, SortIconColors colors);